테스트 에러 해결

+++ s-with-me-back/src/main/java/com/swithme/domain/myBook/MyBook.java
@@ -2,7 +2,6 @@ package com.swithme.domain.myBook;
 
 import com.swithme.domain.book.Book;
 import com.swithme.domain.folder.Folder;
-import com.swithme.domain.folder.FolderRepository;
 import com.swithme.web.dto.MyBookUpdateRequestDto;
 import lombok.Builder;
 import lombok.Getter;
@@ -33,7 +32,7 @@ public class MyBook {
 
 
     @Builder
-    public MyBook(Folder folder, Book book, short lastProblemId){
+    public MyBook(Folder folder, Book book, int lastProblemId){
         this.folder = folder;
         this.book = book;
         this.lastProblemId = lastProblemId;
